+++ src/lib/components/module/ModuleSourceCode.tsx
@@ -11,8 +11,6 @@ import {
   Heading,
   Text,
 } from "@chakra-ui/react";
-import type { Monaco } from "@monaco-editor/react";
-import MonacoEditor from "@monaco-editor/react";
 
 import { AppLink } from "../AppLink";
 import { CopyButton } from "../copy";
@@ -22,16 +20,7 @@ import type { MoveVerifyInfoResponse } from "lib/services/types";
 import { MoveVerifyStatus } from "lib/types";
 import type { Nullish } from "lib/types";
 import { formatUTC } from "lib/utils";
-
-import { moveLanguageConfig, moveTokenProvider } from "./moveSyntax";
-
-const loadMoveSyntax = (monaco: Monaco) => {
-  monaco.languages.register({ id: "move" });
-  monaco.languages.onLanguage("move", () => {
-    monaco.languages.setMonarchTokensProvider("move", moveTokenProvider);
-    monaco.languages.setLanguageConfiguration("move", moveLanguageConfig);
-  });
-};
+import { Editor } from "../editor/Editor";
 
 interface ModuleSourceCodeProps {
   verificationData: Nullish<MoveVerifyInfoResponse>;
@@ -120,14 +109,7 @@ export const ModuleSourceCode = ({
             borderColor="gray.700"
             borderRadius="8px"
           >
-            <MonacoEditor
-              height={400}
-              language="move"
-              theme="vs-dark"
-              beforeMount={loadMoveSyntax}
-              value={verificationData.source}
-              options={{ readOnly: true, scrollBeyondLastLine: false }}
-            />
+            <Editor value={verificationData.source} />
           </Box>
         </AccordionPanel>
       </AccordionItem>